sngpl-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
paho-mqtt==1.6.1
python-dotenv==1.0.0
websockets==12.0
slowapi==0.1.9
schedule==1.2.0
email-validator==2.3.0
reportlab==4.0.7
openpyxl==3.1.2
aiosmtplib==3.0.1
jinja2==3.1.2
psycopg2-binary==2.9.9
